[FIX] l10n_ch: adapt bank account creation in reconciliation
If the bank account number is a postal one, search differently
to see if the account already exists.  As for a postal account,
the same number can be used by multiple partners, changed the
account to include the name of the partner, but for this
we need to alter the way the account is searched/created.

closes odoo/odoo#47365

X-original-commit: b95b9d6db2af51644f640142d5a0b9ca72eac52f
Signed-off-by: Josse Colpaert <jco@openerp.com>