[FIX] web: FieldMany2ManyTags: failing quick create record
The FieldMany2One allows to quick create record (name_create).
However, sometimes, required fields on the co-model make the
name_create crash. In this case, the FieldMany2one opens a form
view in a dialog to let the user specify those fields, and create
a new record with the 'create' method. This was working fine.

The FieldMany2ManyTags embeds a FieldMany2One to let the user
search for and select or (quick) create tags. In this case, failing
quick creation didn't fallback on the dialog form view, but
displayed a server error instead. This commit fixes this issue.

Bug introduced by b0092039e48d0644634d6c1d7d533126e06ecec2

Task 2192754

closes odoo/odoo#47183

X-original-commit: 5d291da39bd542d31caa45e40ec38db6eceff8af
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>