[IMP] website_sale: make "Image full" a global option
Previously, the "Image full" option was on a per product basis. This is
very strange, because in most cases you either want all of your product
images to take up the entire card space, or none of them. This commit
makes the option into a global option for the shop page, so it can be
toggled for all products at once.

task-1981170
Part of odoo/odoo#47574