[FIX] l10n_latam_invoice_document: check of account_move type
The `type` field was renamed `move_type` in
ef8ed61792d45f9e17146a08674be6ec9d9eed62 but a forward-port re-added a
use of the old `type` name, which breaks if an account.move is
unlinked while l10n_latam_invoice_document is unlinked (and there are
related moves to update).

Instead of renaming the field, use the proper utility method for the
category (purchase) provided by account.move.

closes odoo/odoo#47301

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>