[FIX] web: introduce parser for `many2one_reference`
Follow-up of e283aa9

The parser of `many2one_reference` was not introduced when the new field
type was created and then later when the formatter was added, which
leads to errors when the web client is trying to set the value of such
fields.

Since the value of this field is an ID, it is parsed as an integer.

closes odoo/odoo#47330

X-original-commit: db1e16bc041e1d892fd1f3ac5b694e12496a32f9
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>