[FIX] purchase: invoice status
1. Create a product A with 'Control Policy' set to 'On received
   quantities'
2. Create a PO for 10 units
   => status is `no`
3. Validate the PO
   => status is `no`
4. Receive 5 units
   => status is `to invoice`
5. Invoice 5 units
   => status is `invoiced`
6. Receive the remaining quantity
   => status is `to invoice`
7. Invoice the remaining quantity
   => status is `invoiced`

There is an inconsistency between steps 3 and 5: there is pending
material to receive, but the status is different.

This is due to the fact the `order.invoice_ids` is checked, while it
shouldn't affect the computation of the status.

The status should be `no` in steps 3 and 5.

opw-91fbc4c

closes odoo/odoo#47802

X-original-commit: 15330712d2ea3c16c57e01c236c783a6febd1376
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>