[FIX] stock: remove hover effect on popover
The hover effect on popover is the following: the popup is opened when
passing the mouse cursor on the icon and closed once it leave the icon.
The issue is if the popup has a button inside, clicking the button
freeze the popup. It cannot be closed unless the entire page is
refreshed

Task : 2199078

closes odoo/odoo#47408

X-original-commit: ed07131a11cd993be1038711316a3f048f8b0c5f
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>